Refactor Manifest Function
